libtracker-miner: Use separate GFileInfo for dest file in move operations
In move operations, using the same GFileInfo breaks further down if the
information there is to be guessed out of the filename differs (i.e.
file moving from hidden to non-hidden, or the other way around), this
may result in file events not being properly sent for this kind of changes.

Generate a distinct GFileInfo for these situations, so the checks on the
dest file come out correct, and result in the file being properly guessed
as created/deleted from the Tracker perspective.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/196