libtracker-miner: Always return task array on TrackerSparqlBuffer
This used to be the behavior (including on errors) as we want to tag
file(s) causing this as failed, in order to ease bug reports. This
behavior broke in commit 7336faa675e as it started returning either
an error or the task array.

Always return both things here, so we can keep the previous behavior.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/201
Related: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/199