libtracker-miner: Plug possible leak on error paths
Coverity seems really confused about item_add_or_update(), thinking
it can unref the given file argument, while it just holds a temporary
extra ref.

However that ref (and the uri string) might leak if fetching a
GFileInfo failed. Avoid the ref altogether, since this is not subject
to reentrancy issues that might make the reference necessary. And
shuffle the uri string creation so it's not leaked there.

CID: #365678