libtracker-miner: Be more specific in tracker-extract remaining files query
Currently, some extractors (mainly, disc-generic for cue files) link their
nie:InformationElement to more than one nfo:FileDataObject (in disc-generic
case, bin/raw/img/etc files pointed by the cue file being extracted).

This has the unintended side effect of adding the corresponding
nfo:FileDataObject in the corresponding metadata graph, which is seen in
the future by tracker-extract-3 as a file that is missing extraction as
it lacks the tracker:extractorHash property.

From there it becomes a little bit awkward, with tracker-extract-3
attempting (and sort of succeeding) in extracting these files despite
no extractor existing for them, just to fail with inconsistent SPARQL:

    Subject `file:///...' is not in domain `nie:DataObject' of property `nie:interpretedAs'

Be more specific in the query performed by tracker-extract-3 to find
out the files missing extraction, in order to skip these files introduced
in the graph just for cross-referencing purposes. Ensuring there is a
nfo:fileName has somewhat better guarantees that we are dealing with tasks
left by tracker-miner-fs-3.