libtracker-miner: Make GFileInfo a g_autoptr to plug leak
In some paths we get a NULL file info, thus query one in place,
in that case the GFileInfo is being leaked. Since this function
has so many return paths, it is significantly more convenient to
make it a g_autoptr, so it's ensured to be always freed after
exit.