libtracker-miner: Remove unused task pool in TrackerMinerFs
Since the file information extraction in tracker-miner-fs was streamlined
so the GFileInfo would be obtained from the TrackerFileNotifier and pushed
to the upper layers, file extraction became no longer asynchronous and this
task pool became unused (besides, it had a limit of 1 for a long time).

So we were left with the code that did the task pool maintenance, but no
elements were ever added to it. It looks like this code can be simply
peeled off.