libtracker-miner: Avoid stopping during flush if possible
Due to the way we asked the store to create the file content URNs
for us, we fairly often had to stop all the machinery while there was
a SPARQL batch being inserted, since we need information on files
that are in the batch being executed before proceeding with files
being processed now (e.g. a nfo:belongsToContainer relation).

Since the file content URNs are stable and quick to fetch from GIO,
we no longer need to wait for anything here. File processing is
able to continue now despite these "dependency" files being in a
batch that is being inserted.

Still, avoid to have more than one batch in flight, if the SPARQL
buffer gets full again before the prior batch finished insertion,
processing will stop until that happens. This still favors memory
usage over parallellization, if batches are built up quicker than
they are inserted in the database, there will always be worth 2 of
them in flight.