libtracker-miner: Notify created/updated files soon
Let the files found during crawling (either created or updated, and
we already know that by merging the information with database queries)
be notified before waiting for the whole index root to be finished.

This allows early handling from the TrackerMinerFS side, and some way
to let it control the "high water" hint if there are too many files
being queued. Otherwise all files found are added at once, and the
"high water" hint becomes moot.