tracker-miner-fs: Raise batch size
Now that crawling is throttled by the amount of items left to process
and the main culprits of high memory usage on large filesystems are
gone, we can raise the batch size a bit. We can definitely afford a
couple extra megabytes in memory now, so raise the batch size to also
optimize the throughput.