libtracker-miners-common: Use better stable filesystem identifiers
Our new stable identifiers come in the `urn:fileid:$FS_ID:$INODE/$SUFFIX`
format, where $FS_ID is a stable identifier for the filesystem. We
currently use G_FILE_ATTRIBUTE_ID_FILESYSTEM for that, but that may differ
in some circumstances, e.g. removable mounts inserted in different order.

Try harder at obtaining a stable identifier for the filesystem, that will
not change on these runtime conditions. We prefer identifiers in this
order:

- If the mount entry node is an actual partition (e.g. /dev/sda3), look
  up the filesystem UUID with blkid
- If the mount entry points to a non-physical partition (e.g.
  /dev/mapper/luks-$UUID, or $HOST:$FOLDER with NFS), the mount entry
  device name is used.
- If none of these are found (e.g. tmpfs), we still resort to
  G_FILE_ATTRIBUTE_ID_FILESYSTEM.

These identifiers are cached for all available mount entries in mtab
for fast lookups, and are updated on mount entry changes.