libtracker-miner: Make FANotify monitor a subclass of GLib monitor
And fallback into GLib for the situations that FANotify cannot set
up a mark despite being detected at runtime.

Unfortunately, a growingly common situation for this are btrfs
subvolumes. These are currently problematic with FANotify due to
filesystem ID mismatches, thus EXDEV is currently returned. See
https://lore.kernel.org/all/CAOQ4uxjt2+BECu60aisZs_D6pgHK3VbBjju6hzmvG=Ls88UqUA@mail.gmail.com/
for a description of the problem.

Hopefully this situation will be resolved at some point, in the
mean time, transparently fallback to the old situation for users
of btrfs.

Closes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/212