libtracker-miner: Handle FAN_DELETE[_SELF] being emitted separately on dirs
If there is a monitor on a dir and its parent, fanotify used to be able to
coalesce the event so FAN_DELETE and FAN_DELETE_SELF would be both set on
an uniquely sent event.

This seems no longer the case on more recent linux versions, where these
generate 2 separate events, first FAN_DELETE_SELF for the folder being
deleted, and then FAN_DELETE from the parent dir to notify of the parent
folder structure change. This double DELETE event emission is inconsistent
with TrackerMonitor behavior and causes test failures.

Handle these 2 events, by caching both and being able to merge them, but
flushing on the one that is expected last. Fixes tracker-file-notifier-test
on recent linux versions (locally, 5.17.0).