tracker-extract: Initialize cache for stable content URNs at startup
This cache initialization may incur in unexpected ioctls to probe
floppy/cdrom devices, and is nowadays left up to the first extractor
thread that queries a content identifier. This may result in seccomp
issues.

Perform this initialization on the main thread early during initialization
so the extractor threads find a populated cache when querying content
identifiers.

Fixes: https://gitlab.gnome.org/GNOME/tracker/-/issues/355