tracker-miner-fs: Always delete graph nie:InformationElement on create/update
There are some situations where the file monitors cannot distinguish between
a file being created where none existed before, or a file newly created
replacing a previously existing file.

Treat all create/update events the same WRT trimming the previously
existing nie:InformationElements, in order to ensure these updates that
pass as creates also result in the file being reindexed by the
metadata extractor. This only applies to files that would have metadata
extracted.

While at it, simplify the SPARQL and move the code so that it is not
scattered across the function.