libtracker-miner: Fix TrackerFileNotifier cancellation paths
Currently, the code paths handling cancellation of a TrackerCrawler async
request attempt to still emit ::directory-finished and update stats. However
this causes invalid memory accesses on finalization paths, since the
TrackerFileNotifier has been already disposed at the time the async callback
is handled. This may result in crashes.

To fix this, avoid all TrackerFileNotifier manipulation in the cancellation
paths of the async callback, the callback is meant to execute and bail out
cleanly now in these situations. This makes it a responsibility of the
g_cancellable_cancel() caller to instruct the TrackerFileNotifier how to
proceed.

Other callers, like the paths handling unmount and deletion of the folder
being crawled, have also been updated to ensure TrackerFileNotifier
resumes operation correctly after those situations.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/223