libtracker-miner: Ensure timer is started when processing elements
Currently, the GTimer used to log the time spent processing is used
in a broken manner. It is only started once when the miner is started,
and every later batch of processing (e.g. monitor events) just stops
the timer and prints the elapsed time, which simply extends the elapsed
time based on the fixed start time, and info logging about our elapsed
time are forever based on the moment the miner was started.

Since the places where the timer is started currently don't apply to
monitor events after the miner is running, ensure the timer is started
when events do arrive in a state of idleness.

This ensures our elapsed time always reports the time spent in processing
the batch of events, until the miner becomes idle again.