cli: Specify graphs in "tracker3 search" queries
Since the queries did not specify the graph, there is some path
explosion due to the nie:dataSource being in multiple graphs,
so all combinations are shown and this is seen as duplicates in
the result lists.

Specify the graphs in all "tracker3 search" queries, so these
redundant object paths are trimmed, and we show unique results.
This also results in a speed up of the queries.

Closes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/229