tracker-extract-libav: Check for all tags also from streams
At least in Ogg audio and video files the tags are actually inside the
streams not in the format context. Check for the tags in suitable streams
and as fallback for generic information check first the audio stream and
then finally also video stream.

Fix some existing style issues in code.