ci: Collect coverage information from all tests
Instead of doing a specific run to get testing coverage, collect the
information of all tests being done. Since some of these will check for
different build options, it's not possible to do that from a single run.

This way we will get accurate information for all code that is being
actually exercised by the tests.

In consequence, using eatmydata to run the tests is no longer ok,
since there is the possibility that it also eats the coverage reports
that are generated during test runs.

Also, since running tracker-extract in tests with seccomp enabled
prevents coverage information from being written for the metadata
extractors, this also means we now disable the seccomp jail on all
tests (as opposed to just the coverage target).