tracker-extract: Use readonly connection to extract matching CUE sheets
We used to iterate all files in the same folder than the audio file to
look up CUE sheet files, but this may be overly expensive on very
flat folder hierarchies where the audio file shares folder with other
thousands of elements.

Since we have this information indexed already, we can use a readonly
connection to query for these CUE sheet files. It is very likely that
these files are just indexed as nfo:FileDataObject, so we have to look
up the file extension to determine cue files, since mimetype detection
relies on the file extension too, the check is equivalent.

And anyways, keep a fast-path lookup for CUE files that have the same
filename but just different extension. Partly to keep tests happy,
partly to also avoid possibly iterating over thousands of CUE files.

Closes: https://gitlab.gnome.org/GNOME/tracker/-/issues/378