seccomp: Disallow fchown
This is needed by SQLite on some circumstances, but these mostly
apply to databases being opened with other users. This is something
that happens on CI, but is not expected to happen in real circumstances.

Anyhow, SQLite does not check for fchown return value, so just error
out softly if that happens.