libtracker-miners-common: Include btrfs subvolume in stable identifiers
In the case of a btrfs partition with multiple subvolumes, we will
create the stable identifiers based on the mountpoint device name,
this falls short if multiple mountpoint entries exist for the same
partition and different subvolumes.

Make the subvolume name part of the stable identifier, so that
we can distinguish between subvolumes of a single partition.