tracker-miner-fs: Add minimum set of file attributes
When querying the directory in place, we are missing the
basic set of attributes. This triggers critical warnings in
the tracker-file-notifier-test suite on newer glib, when
querying non-added properties:

not ok /libtracker-miner/file-notifier/changes-remove-ignore - GLib-GIO-FATAL-CRITICAL: GFileInfo created without standard::is-hidden

Ensure the basic set here, and make the test suite pass
in time for 3.5.0.