Removing [Install] section from tracker-extract-3.service
Having WantedBy=default.target was triggering the service before the session
environment variable are set in gnome-session.

This service relies on gvfs-daemon.service and by starting it before the
environment variables are set, some important variables like KRB5CCNAME are
missing.

This means that the daemon won't have access to it when trying to
mount shared folders through the network, creating problems like requesting
credentials when trying to gio mount a samba share, despite having a valid
kerberos ticket on the machine.

Since this is a helper service that is meant to be controlled by
"tracker-miner-fs", the install section shouldn't exist, as it allows
the service to enabled, meaning that its execution would be
controlled by systemd.