tracker-miner-fs: Correctly wait for TrackerEndpoint thread initialization
Do things as per GCond documentation. Namely, initialize the wait condition
within the mutually exclusive area, as doing it outside may result in the
main thread going forward ahead of time.

This could result in situations that the endpoint thread is stuck on
initialization. Although it was mostly noticeable in CI, it looks like
a low-chance heisenbug that could make the endpoint not fully responsive.

This make flaky CI much less flaky.

Fixes: 8d6508d9623 ("tracker-miner-fs: Delegate TrackerEndpoint to a separate thread")