tracker-extract: Add missing break to remove a single element
The extractor process listens to notifier events in order to a)
know there's missing elements, in order to build its small cache
(nowadays: 200) of elements to extract, or b) notice files being
deleted, which might be part of the currently cached items.

The loop doing the latter was somewhat broken, trying to remove
elements from a GQueue while iterating it. Since cached elements
must be just once in the list at most, this is just missing a
break.

Closes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/265