tracker-extract: Delete files with unknown graphs
From the time they are processed to the time their content is extracted
by tracker-extract, there is a window of opportunity that the file might
change substantially while its nfo:FileDataObject is already present in
a content graph.

So this file would exist and be on a content graph, thus be a candidate
for tracker-extract-3. Extraction would be attempted and failed, and
deletion of the extractor hash would be attempted. The only thing that
might not quite work is that the file mimetype no longer corresponds to
any extractor hash, so this deletion might implicitly fail and do nothing,
and the file be left there for it to trigger neverending activity from
tracker-extract-3 in this fruitless loop.

Handle this situation where the hash is unknown, resorting to the
deletion of the nfo:FileDataObject from the content graph, and fixing
this situation.