tracker-extract: Trigger ::items-available anytime new items are received
We may receive event notifications at the time the extractor finished
all its cached items and is busy executing the corresponding update batch.
If that happened, the decorator would not be inactive yet, but the upper
layers would have already called tracker_decorator_next() for the "last"
time.

In order to keep the extractor moving in this situation, we should emit
::items-available right away, so that the extractor can resume extraction
while the batch finishes execution.

Fixes: 358faa359 ("tracker-extract: Make TrackerDecorator synchronous")

Closes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/268