tracker-extract: Regain performance at query to get unextracted items
In the tracker-extract query rework towards TrackerSparqlStatement and
static SPARQL queries built in GResources, the query to fetch unextracted
queries changed from a handcrafted version like:

  SELECT ... {
    { GRAPH ... { ?u a nfo:FileDataObject } } UNION
    { GRAPH ... { ?u a nfo:FileDataObject } } UNION
    ...
  }

That accessed each of the individual graphs in the fastest way possible,
ensuring the high-priority graphs get sorted first. This changed to a more
generic format in order to avoid in-place query string building:

  SELECT ... {
    { GRAPH ?g { ?u a nfo:FileDataObject } . FILTER (STR (?g) = ...) } UNION
    ...
  }

That basically results in, for each graph, querying all graphs in order
to filtering the results down to the desired one.

While this works, it multiplies the internal amount of churning and is
thus not great at performance. Since we query remaining elements in rather
small batches, the query time can get prohibitely high on a high enough
amount of indexed elementsn.

In order to improve on this, rework the query so that we get the best of
both worlds, in exchange of some parameter ugliness. The reworked query
works by querying everything in two blocks (for high and regular priority
graphs), for all graphs in each, but controlling the limit of each of
those subqueries:

  SELECT ... {
    {
      { SELECT ... { GRAPH ... { ?u a nfo:FileDataObject }  } LIMIT ~high } UNION
      ...
    } UNION {
      { SELECT ... { GRAPH ... { ?u a nfo:FileDataObject }  } LIMIT ~low } UNION
      ...
    }
  }

So high priority graphs get an unconstrained (-1) limit in the first block and
a no-op (0) limit in the second, and regular priority graphs get the opposite. This
results in direct graph access in the query without resolving ?g, no hand
crafting required, and as fast and efficient in memory as the former query.

The slight ugliness is that we have to specify, for each graph, the limits
for the high and regular priority blocks. Such is life for the sake of
performance.

Fixes: 1afc8bc05e ("tracker-extract: Query remaining items through statement from a GResource")