tracker-miner-fs: Only stash direct children of changed folders
We do a equals-or-descendant check to find out if we are able to notify
about a file immediately (changed or not) or it should wait for the
parent directory to be crawled.

This makes us keep file information for files not directly related to
the change (e.g. files contained in unchanged children directories of
the changed folder).

We can perform a equals-or-parent check, so that we only keep direct
children of changed folders until crawling of the given folder happens.

Fixes: 2872008f89 ("tracker-miner-fs: Change second-time index behavior")