build: Fix handling of datadir, etc. as absolute paths
There's one place in the meson files where path building was done
incorrectly, which cases breakage when some directories are passed as
absolute paths; for example on Exherbo we use
-Dprefix=/usr/x86_64-pc-linux-gnu -Ddatadir=/usr/share in our
cross-prefix configuration.

The settings for LOCALEDIR and SHAREDIR were built by concatenating the
prefix and specific dir variables with a / in between, which failed with
absolute paths. Switch them to use the / operator, and use the
set_quoted function to avoid needing any string formatting.