tracker-miner-fs: Stop indexing applications
This is nowadays unused (Games/Highscore stopped using tracker-miners
time ago), and the possible complexities with multiple layers
of application directories and desktop files able to override others
from similar directories forever flew over our head, and are now
commonplace with Flatpak.

This is not as simple, and it's done for no gain, sounds like
something that may be removed, applications wanting to inspect
available applications may just use GAppInfoMonitor that deals with
the specific complexities of application databases.