tracker-miner-fs: Handle extractor hash updates specifically
With the indexing changes that happened to optimize second-index
behavior, there was a small oversight. If the extractorHash property
ever changed for folders (say, a change in tracker-miner-files-methods.c
triggering a mass change of extractor hashes), each of these folders
would be treated as "updated", thus crawled and their direct children
checked. With this happening en masse for all folders, it's pretty much
a full reindex and a worst case in cpu/memory usage.

Separate the "hash changed" condition, so that files/directories with
a changed hash are checked/re-extracted individually, but not crawled.
This avoids the memory peak involved in queueing every folder for later
crawling, and the CPU involved in pointlessly crawling each of them
afterwards.