tracker-miner-fs: Always fetch next Fanotify event
In some unlikely circumstances, it seems possible to receive Fanotify
events that are not of the expected type, we do fail then (and in
other "bail out" situations) to iterate forward to the next Fanotify
event in the queue, and are thus possibly left stuck handling the
unexpected event in an infinite loop.

Ensure to move forward to the next event in all circumstances, in
order to tiptoe safely around any of those unexpected events.

A local reproducer was dropping '.git' from ignore-directories-with-content,
indexing a kernel tree, and switching branches in that indexed git tree. In
all the burst of events received, at least one of them triggered these
unexpected paths.