tracker-miner-fs: Avoid content-type extraction in up-to-date checks
Nowadays, we command the TrackerFileNotifier to provide the higher layers
a GFileInfo with enough information to fill in all file-side indexed data.
This is most usually great, to avoid doubly processing of stat() data
(first when checking files did or did not change, then later on during
actual extraction), but the content-type is specifically something
potentially expensive to obtain, and unneeded for up-to-date checks.

Make the content-type data be extracted only during extraction, so that
up-to-date checks may avoid falling through these expensive paths (maybe
pointlessly, if the file did not change after all).