tracker-miner-fs: Do not accept patterns in ignore-directories-with-content
This is not documented as a pattern, and it's in fact quite cumbersome that
it is treated as one, since it does require the full directory contents to
be known beforehand in order to apply this filter, with the subsequent
possible memory increase in having to keep contents of full directories in
memory before being able to tell whether they should be indexed or not.

Disallowing patterns here means we can optimize these paths in terms of
memory usage, by directly stat()ing the few relevant files, thus eliminating
the need for keeping the full directory children list in memory at once.