tracker-miner-fs: Stop using TrackerCrawler and enumerate files directly
The TrackerCrawler is pretty much an old-timey iterator that made it easy
to apply the various filters based on file names and directory contents.
We no longer need to keep all directory children in memory for the latter,
so it is possible now to use GFileEnumerator directly, and keep short
buffers of files being handled to avoid indefinite memory growth.