tracker-extract: Throttle extraction of items while on battery
Historically, the metadata extraction task used to be launched on a
per-file basis from the miner side, so any throttling performed on
the miner side would also affect metadata extraction.

At some point that changed so that metadata extraction became its
own "miner", figuring out the next elements to extract metadata from
by itself. But this did run in parallel to the miner, so any throttling
performed there was still able to slow down metadata extraction by
starving the queue.

But at some point later in time, we changed this relation so the
metadata extraction step waited for the miner to become idle, in
order to serialize indexing into a mostly single-CPU task. But it
is at this point that we made metadata extraction basically unthrottled,
since there is nothing stopping it from going through the items as
fast as it is able to.

We want to get some throttling back on metadata extraction while on
battery though. This is something the tracker-extract-3 process needs
to handle by itself, so add a GSource jump to the main loop in the
extraction process, and make it go at max speed (i.e. g_idle_add)
while on a power source but throttle it (i.e. g_timeout_add) to a
maximum of 100 files per second while on battery.

This drastically reduces CPU usage, while still providing somewhat
reasonable deadlines on common cases (i.e. thousands of files).