tracker-extract: Hint file access pattern to the kernel
Unlike the kernel, we do know what files will the metadata extractor
access, and their order. Hint this access pattern through posix_fadvise(),
in a manner that the next file that will be processed is always hinted
to the kernel with POSIX_FADV_WILLNEED to indicate that it will be read
shortly. Any just processed file is likewise hinted as POSIX_FADV_DONTNEED
to indicate the cached data related to the file may now be dropped.