tracker-miner-fs: Check for file existence looking up mimetype
Since commit fa451b7899 we avoid querying the content type until
we are in need to index the file. This however falls into TOCTOU
issues if the file got deleted at the time of doing this check,
we'd go with a NULL mimetype, and crash afterwards.

Before that commit, we'd just go with the content type from the
given GFileInfo, index the file, and trust that we'd later on get
a corresponding DELETED event to bring things back to harmony.

Now, we can still trust that we'll get that DELETED event, but
we can simply bail out early if no content type was found (meaning
the file does not exist), in order to avoid the crash.