tracker-miner-fs: Drop unused property
The TrackerIndexingTree::root property is pretty much
left to defaults. Whether it still makes sense to support
a single URI scheme or not is a question worth making and
addressing in the future. In the mean time, we can just
rely on default behavior from the TrackerIndexingTree
(i.e. supporting file:///).