cli: Separate URN extraction and printing at "tracker3 info"
Separate "tracker3 info" processing in 2 different loops, one to
get the URNs to extract info from out of the given arguments, and
another to print each of them.

This will make the info printing nicer when using modern APIs for
resource description, and friendlier to using modern features
(e.g. piping to pager).