cli: Add "[🡕]" link close to compressed IRIs in "tracker3 info"
This kind of ANSI escape codes for for hyperlinks is relatively recent,
but supported by VTE/xterm and a number of terminal implementations. Systemd
and others do use it in their CLI outputs.

Now that we have working ontology links, brag about them and add such links
besides compressed IRIs for every class and property, so their documentation
is easily accessible.