tracker-miner-fs: Preserve second event coalescing create+create
Crawling and monitoring a directory is inherently racy, in that content
changes happening around that time may either result in the loss of
events (if monitors are added after crawling) or double events may
happen, found both through monitors and crawling (if monitors are added
before crawling).

Tracker takes the second option, and ignores/coalesces the events, and
in the situation that two CREATE events are coalesced, it would resort
for deleting the second and keeping the first in its place in the queue.

This used to work fine, since the GFileInfo was queried later on, just
in time for SPARQL insertion, so it would contain a recent snapshot of
the file. This however has changed in recent times, where we let the
TrackerFileNotifier produce the GFileInfos that end up inserted. This
made every event already contain their GFileInfo that would end up
emitted, so sticking to the event already in the queue would result
in old information being inserted in the database.

This is something typically very unlikely, and anyways caught on next
miner restart, the file change is belatedly detected, and the file
updated.

However, this is a situation that we run head first in tests, and
sometimes hit, the GFileInfo that ends up inserted has 0 size, so it
ends up ignored by the extractor. This is seen as random timeouts
running the test suite.

Since with the GFileInfo preserving scheme it also makes sense for
create+update events to stick to the second event, change that. But
that only holds up if the update event is not an attributes-only update
in which case we should let both events go through.